:sparkles: api: move SessionMassImportReport model to src
