+++ api/lib/domain/usecases/sessions-mass-import/validate-sessions.js
@@ -1,5 +1,5 @@
 import { Session } from '../../../../src/certification/session/domain/models/Session.js';
-import { SessionMassImportReport } from '../../models/SessionMassImportReport.js';
+import { SessionMassImportReport } from '../../../../src/certification/session/domain/models/SessionMassImportReport.js';
 import { CertificationCandidate } from '../../models/CertificationCandidate.js';
 import bluebird from 'bluebird';
