+++ api/src/certification/session/domain/models/SessionMassImportReport.js
